pyproject.toml
[tool.poetry]
name = "vm-lifecycle"
version = "0.1.0"
description = ""
authors = ["Lorcan Rae <lorcanrae@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.8"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[tool.poetry.scripts]
vmlc = 'vmlc.main:cli'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
